intermediate commit

+++ internal/provider/server_data_source.go
@@ -274,12 +274,12 @@ func serverToDataSourceModel(ctx context.Context, s *client.Server, data *Server
 		Enabled: types.BoolValue(cfg.StatsMonitoring),
 		Types:   typesSet,
 		Thresholds: &ServerAlertsThresholdsModel{
-			CPUCritical:    types.Int64Value(int64(cfg.CPUCritical)),
-			CPUWarning:     types.Int64Value(int64(cfg.CPUWarning)),
-			DiskCritical:   types.Int64Value(int64(cfg.DiskCritical)),
-			DiskWarning:    types.Int64Value(int64(cfg.DiskWarning)),
-			MemoryCritical: types.Int64Value(int64(cfg.MemCritical)),
-			MemoryWarning:  types.Int64Value(int64(cfg.MemWarning)),
+			CPUCritical:    types.Float64Value(cfg.CPUCritical),
+			CPUWarning:     types.Float64Value(cfg.CPUWarning),
+			DiskCritical:   types.Float64Value(cfg.DiskCritical),
+			DiskWarning:    types.Float64Value(cfg.DiskWarning),
+			MemoryCritical: types.Float64Value(cfg.MemCritical),
+			MemoryWarning:  types.Float64Value(cfg.MemWarning),
 		},
 	}
 }
